# improve ImageTool performance