with #1686 supporting scheduled event images, ImageTool has come under my focus.

## explanation

the ImageTool shipped in v4 reads the provided stream to a string builder, then returns the contents of that string. this incurs a huge string allocation at the end that we can't avoid (more details on that later), *but it also incurs an intermediary allocation for the data conversion*, as this conversion happens using `Convert.ToBase64String`, thereby creating one string of just the b64'ed data *and another of the b64'ed data and the format information prepended*, which is then returned; and it incurs another allocation for the data before b64'ing, as it reads the entire stream into an allocated byte array

this new ImageTool uses a pooled array of 3KB to read from the stream, then immediately converts that to b64 and writes to the also pooled output buffer. thus, the only allocations incurred are for the initial pooled arrays, which can be quite substantial (an 8mb image incurs a 16mb `ArrayPool` allocation, since it normalizes array sizes to the next higher power of two - though even if we pooled this ourselves, the allocation would still be ~11mb), and for the inevitable string allocation. this, in turn, drastically reduces allocations per iteration, though CPU time stays in unacceptable ranges (see performance section).

the string allocation mentioned several times is, in fact, required by our serialization infrastructure. there may be ways to adapt our infrastructure to handle this better, but they are not explored in this pull request, thus, we are forced to provide a `System.String` containing image data. while Newtonsoft.Json *does* support `byte[]` serialization into base64, we need to prepend a small amount of plaintext data according to the scheme laid out in [the API docs](https://discord.com/developers/docs/reference#image-data) - `data:image/FORMAT;base64,B64_IMAGE_DATA`, and were we to use `byte[]`, this format information would also get converted to b64.

## performance


| Method     | Mean     | Error    | StdDev   | Ratio | Gen0     | Gen1     | Gen2     | Allocated | Alloc Ratio |
|----------- |---------:|---------:|---------:|------:|---------:|---------:|---------:|----------:|------------:|
| V4-STABLE | 14.75 ms | 0.294 ms | 0.245 ms |  1.00 | 828.1250 | 828.1250 | 828.1250 |  65.01 MB |        1.00 |
| V5-PR1687      | 13.90 ms | 0.072 ms | 0.060 ms |  0.94 | 953.1250 | 953.1250 | 953.1250 |  19.27 MB |        0.30 |

data obtained from a PNG with a size of 7,574,458 bytes.

here, it is noteworthy that some allocations in v5-PR1687 are pooled, meaning the allocations per iteration actually only amount to 16MB.
conversion time has also improved, but only marginally - in fact, ~95% of the total iteration time is spent purely on allocating and populating the string.

these measures further jump up considerably when sending the data as HTTP request, to up to ~85MB for v4-stable and ~27MB for v5-pr1687, which is unavoidable with our current rest/serialization infrastructure.